refactor(sync): isolate mirror state and formatting
